fix(pin_integrity): changed route and added openapi
